feat: 시간에 예약이 있을 시, 시간 삭제 방지 기능 추가

+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -3,7 +3,9 @@ package roomescape.service;
 import org.springframework.stereotype.Service;
 import roomescape.dto.ReservationTimeRequest;
 import roomescape.dto.ReservationTimeResponse;
+import roomescape.model.Reservation;
 import roomescape.model.ReservationTime;
+import roomescape.repository.ReservationRepository;
 import roomescape.repository.ReservationTimeRepository;
 
 import java.util.List;
@@ -12,9 +14,11 @@ import java.util.List;
 public class ReservationTimeService {
 
     private final ReservationTimeRepository reservationTimeRepository;
+    private final ReservationRepository reservationRepository;
 
-    public ReservationTimeService(ReservationTimeRepository reservationTimeRepository) {
+    public ReservationTimeService(ReservationTimeRepository reservationTimeRepository, ReservationRepository reservationRepository) {
         this.reservationTimeRepository = reservationTimeRepository;
+        this.reservationRepository = reservationRepository;
     }
 
     public List<ReservationTimeResponse> getAllReservationTimes() {
@@ -27,10 +31,17 @@ public class ReservationTimeService {
     public ReservationTimeResponse addReservationTime(ReservationTimeRequest reservationTimeRequest) {
         ReservationTime reservationTime = ReservationTimeRequest.toReservationTime(reservationTimeRequest);
         ReservationTime savedReservationTime = reservationTimeRepository.save(reservationTime);
+
         return ReservationTimeResponse.from(savedReservationTime);
     }
 
     public void deleteReservationTimeById(Long id) {
+        List<Reservation> reservations = reservationRepository.findAllByReservationTimeId(id);
+
+        if(!reservations.isEmpty()) {
+            throw new IllegalArgumentException("해당 시간에 예약이 존재합니다.");
+        }
+
         reservationTimeRepository.deleteById(id);
     }
 }
